% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gr.r
\name{eval_gr}
\alias{eval_gr}
\title{Computation of growth curves}
\usage{
eval_gr(dat, curve_type, equation_type)
}
\arguments{
\item{dat}{data.frame. See}

\item{curve_type}{character string. It determines which curve to use: 'logistic' (default),
'schumacher', 'gompertz', 'monomolecular', 'arctangent', 'hyperbolic' or 'user'.}

\item{equation_type}{character string. It can be equal to 'td', 'rate' or 'ti',
indicating whether to calculate growth time-dependent size, rate parameter or
time-independent size, respectively.}
}
\value{
vector with length equal to the number of rows in \code{dat}, containing
either the growth rate parameter or size.
}
\description{
Numerical evaluation of growth curves
}
\details{
The user is not supposed to call this function directly.
Use \code{\link{growth_rate}}, \code{\link{ti_size}} or \code{\link{td_size}} instead.
}
\examples{
dat <- data.frame(y1 = 10, y2 = 43, tdiff = 10, max_y = 80, t = 25, k = .025, offset = -.1)
eval_gr(dat, "logistic", "td")
eval_gr(dat, "logistic", "rate")
eval_gr(dat, "logistic", "ti")

}
