% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_optim.R
\name{fit_optim}
\alias{fit_optim}
\title{Title}
\usage{
fit_optim(dat, fo, coef, method = "Nelder-Mead")
}
\arguments{
\item{dat}{data.frame}

\item{fo}{formula object}

\item{coef}{named numeric vector with initial values of the coefficients.}

\item{method}{character string corresponding to the \code{method} option in the \code{optim} function.}
}
\value{
A \code{nls} object.
}
\description{
Title
}
\examples{
dat <- data.frame(x1 = -10:100, x2 = runif(111), y = rnorm(111))
fo <- y ~ a * x1 + log(b * x2)
coef <- c(a = 2, b = .2)
fit_optim(dat, fo, coef)
}
