% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_arctangent.r
\name{gr_arctangent}
\alias{gr_arctangent}
\title{Arc-tangent equations for growth}
\usage{
gr_arctangent(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_arctangent} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_arctangent("td"))
print(gr_arctangent("rate"))
print(gr_arctangent("ti"))

}
