% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_arctangent_exp.R
\name{gr_arctangent_exp}
\alias{gr_arctangent_exp}
\title{Arc-tangent exponential equations for growth}
\usage{
gr_arctangent_exp(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_arctangent_exp} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_arctangent_exp("td"))
print(gr_arctangent_exp("rate"))
print(gr_arctangent_exp("ti"))

}
