% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_gompertz.r
\name{gr_gompertz}
\alias{gr_gompertz}
\title{Gompertz equations for growth}
\usage{
gr_gompertz(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_gompertz} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_gompertz("td"))
print(gr_gompertz("rate"))
print(gr_gompertz("ti"))

}
