% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_hyperbolic.r
\name{gr_hyperbolic}
\alias{gr_hyperbolic}
\title{Hyperbolic equations for growth}
\usage{
gr_hyperbolic(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_hyperbolic} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_hyperbolic("td"))
print(gr_hyperbolic("rate"))
print(gr_hyperbolic("ti"))

}
