% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_logistic.r
\name{gr_logistic}
\alias{gr_logistic}
\title{Logistic equations for growth}
\usage{
gr_logistic(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_logistic} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_logistic("td"))
print(gr_logistic("rate"))
print(gr_logistic("ti"))

}
