% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_monomolecular.r
\name{gr_monomolecular}
\alias{gr_monomolecular}
\title{Monomolecular equations for growth}
\usage{
gr_monomolecular(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_monomolecular} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_monomolecular("td"))
print(gr_monomolecular("rate"))
print(gr_monomolecular("ti"))

}
