% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_rational.R
\name{gr_rational}
\alias{gr_rational}
\title{Rational equation for growth}
\usage{
gr_rational(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_rational} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_rational("td"))
print(gr_rational("rate"))
print(gr_rational("ti"))

}
