% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_schumacher.r
\name{gr_schumacher}
\alias{gr_schumacher}
\title{Schumacher equation for growth}
\usage{
gr_schumacher(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which expression will be
returned, namely growth rate ("rate"), time-independent ("ti")
or time-dependent ("td") growth.}
}
\value{
R string containing the selected equation.
}
\description{
\code{gr_schumacher} returns a character string containing the expression for
the selected equation type.
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_schumacher("td"))
print(gr_schumacher("rate"))
print(gr_schumacher("ti"))

}
