% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Rtigre}
\alias{predict.Rtigre}
\title{Predict method for nls2 objects}
\usage{
\method{predict}{Rtigre}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class "Rtigre".}

\item{newdata}{A data frame in which to look for variables with which to predict.}

\item{...}{Not used.}
}
\value{
A numeric vector of predictions.
}
\description{
Predicts from a fitted Rtigre model. If the 'log.transf' attribute is TRUE, the predictions are exponentiated.
}
\details{
To undo the log-transformation implemented in the linear regression we must exponentiate the predicted values
as follows:

\deqn{
  \text{Predicted Y} = e^{\hat{y} + \frac{\hat{\sigma}^2}{2}}
}
where \eqn{\hat{y}} is the predicted value on the log scale and \eqn{\hat{\sigma}^2}
is the variance of the model residuals.
}
