% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtigre.R
\name{Rtigre}
\alias{Rtigre}
\title{Create a new Rtigre object}
\usage{
Rtigre(nls_obj, log_transf = FALSE)
}
\arguments{
\item{nls_obj}{An object of class "nls".}

\item{log_transf}{A logical value indicating whether predictions should be exponentiated.}
}
\value{
An object of class "Rtigre".
}
\description{
Creates a new object of class "Rtigre" from an existing nls object.
}
