% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_gr.r
\name{string_gr}
\alias{string_gr}
\title{Choice of growth curve}
\usage{
string_gr(curve_type = "logistic", equation_type = "rate")
}
\arguments{
\item{curve_type}{string. It indicates which curve type to provide. Possible
values are: "logistic", "schumacher", "gompertz", "monomolecular", "arctangent",
"hyperbolic" or "user".}

\item{equation_type}{string, can be equal to "rate", "ti" or "td". See Details.}
}
\value{
a character string containing an expression for the growth rate parameter
(\code{equation_type="rate"}),
age/time-independent (\code{equation_type="ti"})
or age/time-dependent (\code{equation_type="td"}) growth functions.
For a detailed definition of those expressions, see accompanying vignette.
}
\description{
\code{string_gr} returns a character string for the selected curve type.
}
\examples{
string_gr("hyperbolic", "td")

}
