% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeIFN.R
\docType{data}
\name{treesIFN}
\alias{treesIFN}
\title{Tree data from the Spanish Inventario Forestal Nacional 2 and 3}
\format{
A data.frame with five columns containing a plot identifier "idplot",
tree species name and the dbh for the Inventarios 2 (dbh2) and 3 (dbh3) in cm.
}
\source{
https://www.miteco.gob.es/es/biodiversidad/temas/inventarios-nacionales/inventario-forestal-nacional.html
}
\usage{
data(treesIFN)
}
\description{
DBH adult tree data from random plots of the Spanish Inventario Forestal Nacional.
Only two species have been selected: Pinus halepensis and Quercus ilex.
}
\examples{
data(treesIFN)
}
\keyword{datasets}
