% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_gr.r
\name{eval_gr}
\alias{eval_gr}
\title{Computation of growth curves}
\usage{
eval_gr(dat, curve_type, equation_type)
}
\arguments{
\item{dat}{\code{data.frame} containing data to evaluate the curves.}

\item{curve_type}{\code{character} string. It determines which curve to use: 'logistic' (default),
'schumacher', 'gompertz', 'monomolecular', 'arctangent', 'hyperbolic', 'arctangent_exp', 'rational' or 'user'.}

\item{equation_type}{\code{character} string. It should be equal to 'td', 'rate' or 'ti',
indicating whether to calculate growth time-dependent size, rate parameter or
time-independent size, respectively.}
}
\value{
vector with length equal to the number of rows in \code{dat}, containing
either the growth rate parameter or size.
}
\description{
\code{eval_gr} Numerical evaluation of growth curves
}
\details{
Depending on the value of the \code{equation_type} parameter \code{dat}
must have the following columns: "t", "k", "offset" and "max_y" (\code{equation_type = "td"}),
"y1", "y2", "tdiff" and "max_y" (\code{equation_type = "rate"}) and
"y1", "k", "tdiff" and "max_y" (\code{equation_type = "ti"}).

Notice that subindices 1 and 2 refer to time points, so that
"y1" and "y2" are the quantities of interests at \eqn{t_{1}} and \eqn{t_{2}}, respectively.
See Vignettes for details.

Although exported, the user is not supposed to call this function directly.
}
\examples{
dat <- data.frame(y1 = 10, y2 = 43, tdiff = 10, max_y = 80, t = 25, k = .025, offset = -.1)
eval_gr(dat, "logistic", "td")
eval_gr(dat, "logistic", "rate")
eval_gr(dat, "logistic", "ti")

}
