% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_fits.R
\name{evaluate_fits}
\alias{evaluate_fits}
\title{Title}
\usage{
evaluate_fits(
  df,
  fo,
  log_transf = FALSE,
  train_proportion = 0.8,
  num_simulations = 100,
  positive_rate = FALSE,
  k_param = NULL,
  algorithm = "nlsLM"
)
}
\arguments{
\item{algorithm}{}
}
\description{
Title
}
