% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.r
\name{fit_growth}
\alias{fit_growth}
\title{Non-linear fit to a time-independent growth curve.}
\usage{
fit_growth(
  dat,
  fo,
  curve_type = "logistic",
  log_transf = FALSE,
  positive_rate = FALSE,
  k_param = NULL,
  algorithm = "nlsLM",
  verbose = T
)
}
\arguments{
\item{dat}{\code{data.frame} with columns containing the dependent variable and the set
of predictors to be used in the fit.}

\item{fo}{\code{formula} describing the right-hand-side of the dependence of the
growth rate on the predictors.}

\item{curve_type}{character indicating the type of growth curve to be used.}

\item{log_transf}{logical. If TRUE, a log-transformation will be applied to
the dependent variable.}

\item{positive_rate}{logical parameter to ensure that growth rate 'k' is always positive.
If TRUE, growth rate 'k' is approximated by a 'softplus' function. In that case, 'k_param'
corresponds to a 'softplus' parameter such that larger 'k_param' values make the elbow
of the "softplus" curve more pronounced.}

\item{k_param}{numeric indicating the value of the softplus transformation parameter.
If \code{positive_rate = FALSE} or not set, \code{k_param} is not evaluated.
If \code{positive_rate = TRUE} and not set, it will be assumed  \code{k_param=1}.}

\item{algorithm}{character, the algorithm to be used. See 'Details'.}

\item{verbose}{logical. If TRUE, information on the progress of the regression is produced.}
}
\value{
a \code{nls} object. See \link[stats]{nls} for details. A 'log_transf' logical attribute is
added to the \code{nls} object and is set to the value of the \code{log_transf} parameter.
}
\description{
Non-linear least-squares fit of growth data to a time-independent growth curve.
}
\details{
The algorithm first fits an explicit expression of the growth rate as a function of the
explanatory variables. The resulting coefficients will be the starting values for the
final non-linear regression.
If \code{positive_rate = FALSE} a softplus transformation is performed to ensure that
the growth rate 'k' is always positive.
The default non-linear least-squares algorithm is "nlsLM", which is implemented in package
\code{minpack}. Other allowed values are "nls", which uses the built-in \code{nls}
function and "nlsr" from package \code{nlsr}. The Levenberg-Marquardt in \link[minpack.lm]{nlsLM}
is very fast and stable and should work fine most of the time.
}
