% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.r
\name{fit_growth}
\alias{fit_growth}
\title{Non-linear fit to a time-independent growth curve.}
\usage{
fit_growth(
  dat,
  fo,
  curve_type = "logistic",
  log_transf = FALSE,
  positive_rate = FALSE,
  k_param = NULL,
  algorithm = "nlsLM",
  verbose = T
)
}
\arguments{
\item{dat}{\code{data.frame} with columns containing the dependent variable and the set
of predictors to be used in the fit.}

\item{fo}{\code{formula} describing the right-hand-side of the dependence of the
growth rate on the predictors.}

\item{curve_type}{character indicating the type of growth curve to be used.}

\item{log_transf}{logical. If TRUE, a log-transformation will be applied to
the dependent variable.}

\item{positive_rate}{logical parameter to ensure that growth rate 'k' is always positive.
If TRUE, growth rate 'k' is approximated by a 'softplus' function. In that case, 'k_param'
corresponds to a 'softplus' parameter such that larger 'k_param' values make the elbow
of the "softplus" curve more pronounced.}

\item{k_param}{numeric indicating the value of the softplus transformation parameter.
If \code{positive_rate = FALSE} or not set, \code{k_param} is not evaluated.
If \code{positive_rate = TRUE} and not set, it will be assumed  \code{k_param=1}.}

\item{algorithm}{character, the algorithm to be used. See 'Details'.}

\item{verbose}{logical. If TRUE, information on the progress of the regression is produced.}
}
\value{
a \code{nls} object. See \link[stats]{nls} for details. A 'log_transf' logical attribute is
added to the \code{nls} object and is set to the value of the \code{log_transf} parameter.
}
\description{
Non-linear least-squares fit of growth data to a time-independent growth curve.
}
\details{
The algorithm first fits an explicit expression of the growth rate as a function of the
explanatory variables. The resulting coefficients will be the starting values for the
final non-linear regression.
If \code{positive_rate = FALSE} a softplus transformation is performed to ensure that
the growth rate 'k' is always positive.
The default non-linear least-squares algorithm is "nlsLM", which is implemented in package
\code{minpack}. Other allowed values are "nls", which uses the built-in \code{nls}
function and "nlsr" from package \code{nlsr}. The Levenberg-Marquardt in \link[minpack.lm]{nlsLM}
is very fast and stable and should work fine most of the time.
}
\examples{

## Common parameters. Simple example.
npoints <- 100
tdiff <- 5
t <- sample(10:150, npoints, replace = T)
max_y <- 120

## Fake climatic data.
temp <- rnorm(npoints, mean = 15.4, sd = 1)
prec <- rnorm(npoints, mean = 560, sd = 50)
Intercept <- .1
coef_temp <- -.00061
coef_prec <- -.000052
k <- Intercept + coef_temp*temp + coef_prec*prec + rnorm(npoints)*.01
y1 <- max_y/(1+exp(-(k*t-5)))
y2 <- max_y/(1+exp(-(k*(t+tdiff)-5))) + rnorm(npoints)*.01
dat <- data.frame(tdiff = tdiff, max_y = max_y, y1 = y1, y2 = y2, temp = temp, prec = prec, Intercept = 1)
r <- fit_growth(dat, ~ Intercept + temp + prec, curve_type = "logistic", log_transf = F, positive_rate = F)
print(summary(r))
plot(with(dat, y2-y1), predict(r), pch = 16, cex = .1, log = "xy")
points(c(0.01, 50), c(0.01, 50), type = "l", lwd = 2, col = "red")

## Same data, but forcing growth rate to be strictly positive.
r <- fit_growth(dat, ~ Intercept + temp + prec, curve_type = "logistic", log_transf = F, positive_rate = T)
print(summary(r))
plot(with(dat, y2-y1), predict(r), pch = 16, cex = .1, log = "xy")
points(c(0.01, 50), c(0.01, 50), type = "l", lwd = 2, col = "red")

## Actual Pinus uncinata data from the Spanish Forest Inventories.
data("Punci_IFN")

## Add time difference between second and third Inventory.
Punci_IFN$tdiff <- 10

r <- fit_growth(Punci_IFN, ~ prec + temp, log_transf = F, positive_rate = F)
print(summary(r))
plot(with(Punci_IFN, y2-y1), predict(r), pch = 16, cex = .1, log = "xy")
points(c(0.01, 50), c(0.01, 50), type = "l", lwd = 2, col = "red")

r <- fit_growth(Punci_IFN, ~ prec + temp, log_transf = F, positive_rate = T)
print(summary(r))
plot(with(Punci_IFN, y2-y1), predict(r), pch = 16, cex = .1, log = "xy")
points(c(0.01, 50), c(0.01, 50), type = "l", lwd = 2, col = "red")

r <- fit_growth(Punci_IFN, ~ prec + temp, log_transf = T, positive_rate = T)
print(summary(r))
plot(with(Punci_IFN, log(y2-y1)), predict(r), pch = 16, cex = .1)
points(c(-50, 50), c(-50, 50), type = "l", lwd = 2, col = "red")

# We try all curve types at once and implement a training-test (80\%-20\%) approach.
curves <- Rtigre:::all_curve_types()
curves <- curves[-match("user", curves)]
npoints <- nrow(Punci_IFN)
ntrain <- round(npoints*.8)
nsimu <- 100
fo <- ~ prec + temp

out <- data.frame(matrix(0, nsimu, length(curves)))
names(out) <- curves
for (i in 1:nsimu) {
  j <- sample(npoints)[1:ntrain]
  dat <- Punci_IFN[j, ]
  r <- lapply(curves, function(x) fit_growth(dat, fo, log_transf = T, positive_rate = F, curve_type = x, verbose = F))
  dat <- Punci_IFN[-j, ]
  out[i, ] <- sapply(r, function(x) cor(predict(x, newdata = dat), (dat$y2-dat$y1))^2)
}
library(ggplot2)
library(gridExtra)
plot_list <- list()
for (i in curves) {
  plot_list[[i]] <- ggplot(out, aes_string(x = i)) +
    geom_histogram(breaks = seq(.8, 1.5, by = .02), fill = "skyblue", color = "black") +
    ggtitle(i) +
    theme_minimal()
}
grid.arrange(grobs = plot_list, nrow = 3, ncol = 3)

}
