% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_growth.r
\name{fit_growth}
\alias{fit_growth}
\title{Non-linear fit to a time-independent growth curve.}
\usage{
fit_growth(
  dat,
  fo,
  curve_type = c("logistic", "schumacher", "gompertz", "monomolecular", "arctangent",
    "hyperbolic", "user"),
  sigmoid = T,
  kmax = NULL,
  log_transf = T,
  verbose = T,
  algorithm = "nlsLM"
)
}
\arguments{
\item{dat}{data.frame containing data to be used in the fit.}

\item{fo}{formula describing the right-hand-side of the dependence of the
growth rate on the predictors.}

\item{curve_type}{character. It indicates the type of growth curve to be used.}

\item{sigmoid}{logical. If TRUE, growth rate will be a sigmoid (i.e. logistic).}

\item{kmax}{NULL or numeric. If NULL, its value will be calculated from the
data.}

\item{log_transf}{logical. If TRUE, a log-transformation will be applied to
the diameter increment.}

\item{verbose}{logical. If TRUE, information on the progress of the regression is produced.}

\item{algorithm}{character, the algorithm to be sed. See 'Details'.}
}
\value{
a \code{nls} object. See ?nls for details.
}
\description{
Non-linear least-squares fit of growth data to a time-independent growth curve.
}
\details{
The default algorithm is "nlsLM", which implements the "nlsLM" function in package
\code{minpack}. Other allowed values are "nls", which used the built-in \code{nls}
function and "nlsr" from package \code{nlsr}.
}
\examples{

## Common parameters. Simple example.
tdiff <- 5
t <- seq(1,100,by=tdiff)
max_y <- 120
k <- .1

## Fake climatic data.
temp <- runif(100,18.6,21.3)
prec <- runif(100,359,514)
t <- c(10,20)
intercept <- .02
coef_temp <- .00061
coef_prec <- .000052
k <- intercept+coef_temp*temp+coef_prec*prec+rnorm(length(temp))*.001
y1 <- max_y/(1+exp(-(k*t[1]-2)))
y2 <- max_y/(1+exp(-(k*t[2]-2)))
dat <- data.frame(tdiff=t[2]-t[1],max_y=max_y,y1=y1,y2=y2,temp=temp,prec=prec)


## Same data, but simulating a sigmoid rate.
k <- 2/(1+exp(-(intercept+coef_temp*temp+coef_prec*prec)))+rnorm(length(temp))*.001
y1 <- max_y/(1+exp(-(k*t[1]-2)))
y2 <- max_y/(1+exp(-(k*t[2]-2)))
dat <- data.frame(tdiff=t[2]-t[1],max_y=max_y,y1=y1,y2=y2,temp=temp,prec=prec)


## Actual Pinus uncinata data from the Spanish Forest Inventories.
data("Punci_IFN")

## Add time difference between second and third Inventory.
Punci_IFN$tdiff <- 10

r <- fit_growth(Punci_IFN, ~prec+temp)
summary(r)


}
