% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_optim.R
\name{fit_optim}
\alias{fit_optim}
\title{Title}
\usage{
fit_optim(dat, fo, coef, method = "Nelder-Mead")
}
\arguments{
\item{fo}{}
}
\description{
Title
}
\examples{
dat <- data.frame(x1 = -10:100, x2 = runif(111), y = rnorm(111))
fo <- y ~ a * x1 + log(b * x2)
coef <- c(a = 2, b = .2)
fit_optim(dat, fo, coef)
}
