% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rate.r
\name{fit_rate}
\alias{fit_rate}
\title{Regression on rate parameter}
\usage{
fit_rate(
  dat,
  fo,
  curve_type = "logistic",
  positive_rate = FALSE,
  k_param = NULL
)
}
\arguments{
\item{dat}{\code{data.frame} containing at least four columns named 'y1', 'y2', 'tdiff' and 'max_y'. See Details
and accompanying Vignettes for a description.}

\item{fo}{an object of class "formula"}

\item{curve_type}{character indicating the type of growth curve to be used.}

\item{positive_rate}{see \code{\link{fit_growth}}.}

\item{kmax}{numeric. If NULL, \code{kmax} will be estimated from the data.}
}
\value{
value of rate parameter for each row in 'dat'.
}
\description{
\code{fit_rate} computes a regression of the rate parameters against a set of predictor variables.
obtained at two different times.
}
\details{
Some of the growth equations are taken from Table 6.2 in #' Burkhart and Tomé (2012).
Columns  'y1' and 'y2' in 'dat' correspond to the sizes of the individual at 't1' and 't2', with column 'tdiff'='t2'-'t1'.
Column 'max_y' correspond to maximum size attainable by the individual when time tends to infinite.

The 'positive_rate = TRUE' option allows us to guarantee that \code{k} always remains positive.
Negative \code{k} values may happen when growth curves
are used to calculate growth under conditions much different from the initial ones.

An intercept is not assumed in the regression model. If we seek to include one, add a constant column
to the input data.frame 'dat' with 1's and specify this column in the model 'fo' (see example below).
}
\examples{

# In the examples below we will calculate a time-dependent growth curve and will
# recover the input k-value.

## Common parameters. Simple example.
tdiff <- 5
t <- seq(1, 100, by = tdiff)
max_y <- 120
k <- .1

dat <- data.frame(t = t, max_y = max_y, offset = 2, k = .05, Intercept = 1)
y <- td_size(dat, "logistic") + rnorm(length(t))*.01
plot(t, y, xlab = "Time", ylab = "Size")

dat <- data.frame(tdiff = tdiff, max_y = max_y, y1 = y[-length(t)], y2 = y[-1])

# We add an intercept term.
dat$Intercept <- 1

r1 <- fit_rate(dat, ~ Intercept)

## Fake climatic data.
temp <- runif(100, 18.6, 21.3)
prec <- runif(100, 359, 514)
t <- c(10, 20)
Intercept <- .02
coef_temp <- .00061
coef_prec <- .000052
k <- Intercept + coef_temp*temp + coef_prec*prec + rnorm(length(temp))*.001
y1 <- max_y/(1+exp(-(k*t[1]-2)))
y2 <- max_y/(1+exp(-(k*t[2]-2)))
dat <- data.frame(tdiff = t[2]-t[1], max_y = max_y, y1 = y1, y2 = y2, temp = temp, prec = prec, Intercept = 1)

## Logistic growth.
r2 <- fit_rate(dat, ~ Intercept)
summary(r2)

## A better model.
r3 <- fit_rate(dat, ~ Intercept + temp + prec)
summary(r3)

## Assuming a sigmoid expression for k.
r4 <- fit_rate(dat, ~ Intercept + temp + prec, positive_rate = TRUE)
summary(r4)

## Actual Pinus uncinata data from the Spanish Forest Inventories.
data("Punci_IFN")

## Add time difference between second and third Inventory.
Punci_IFN$tdiff <- 10
Punci_IFN$Intercept <- 1

k1 <- fit_rate(Punci_IFN, ~ Intercept + prec + temp)
k2 <- fit_rate(Punci_IFN, ~ Intercept + prec + temp, positive_rate = TRUE)
summary(k1)
summary(k2)

}
\references{
Burkhart, Harold E., and Margarida Tomé. "Growth functions." In Modeling forest trees and stands,
pp. 111-130. Springer, Dordrecht, 2012.
}
