% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_gompertz.r
\name{gr_gompertz}
\alias{gr_gompertz}
\title{Gompertz equations for growth}
\usage{
gr_gompertz(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which equation will be
calculated, namely growth rate, time-independent or time-dependent growth.}
}
\value{
R expression of equation.
}
\description{
Gompertz equations for growth
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_gompertz("ti"))

}
