% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_hyperbolic.r
\name{gr_hyperbolic}
\alias{gr_hyperbolic}
\title{Hyperbolic equations for growth}
\usage{
gr_hyperbolic(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which equation will be
calculated, namely growth rate, time-independent or time-dependent growth.}
}
\value{
R expression of equation.
}
\description{
Hyperbolic equations for growth
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_monomolecular("ti"))

}
