% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_logistic.r
\name{gr_logistic}
\alias{gr_logistic}
\title{Equations for logistic growth}
\usage{
gr_logistic(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It yields the equation corresponding to growth rate,
time-independent growth or the usual time-dependent logistic curve.}
}
\value{
character string with equation.
}
\description{
Equations for logistic growth
}
\examples{
print(gr_logistic(equation_type = "independent"))
}
