% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_monomolecular.r
\name{gr_monomolecular}
\alias{gr_monomolecular}
\title{Monomolecular equations for growth}
\usage{
gr_monomolecular(equation_type = "rate")
}
\arguments{
\item{equation_type}{character. It indicates which equation will be
calculated, namely growth rate, time-independent or time-dependent growth.}
}
\value{
R expression of equation.
}
\description{
Monomolecular equations for growth
}
\details{
See accompanying Vignette.
}
\examples{

print(gr_monomolecular("td"))
print(gr_monomolecular("rate"))
print(gr_monomolecular("ti"))

}
