% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_gr.r
\name{string_gr}
\alias{string_gr}
\title{Choice of growth curve}
\usage{
string_gr(curve_type = "logistic", equation_type = "rate")
}
\arguments{
\item{curve_type}{string. It indicates which curve type to provide. Possible
values are: "logistic", "schumacher", "gompertz", "monomolecular", "arctangent",
"hyperbolic" or "user".}

\item{equation_type}{}
}
\value{
a character string containing an expression for growth rate parameter,
age/time-independent or age/time-dependent growth functions. For a detailed
definition of those expressions, see accompanying vignette.
}
\description{
Yields a character string for one of seven growth curves.
}
\examples{
string_gr("hyperbolic", "td")

}
